backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
asyncpg==0.30.0
python-dotenv==1.1.1
psycopg2-binary==2.9.9
pydantic[email]==2.11.7
pydantic-settings==2.3.0
python-multipart==0.0.20
jinja2==3.1.5
reportlab==4.1.0
weasyprint==62.3
aiofiles==24.1.0
